test build

+++ cheese/admin/files/scripts/log.js
@@ -29,6 +29,16 @@ function deleteLog(log) {
     });
 }
 
+function deleteAll() {
+    url = "/admin/deleteAll";
+    
+    return new Promise(resolve => {
+        sendGet(url, debug, function(response){
+            resolve(response);
+        });  
+    });
+}
+
 async function buildLogTable() {
     response = await getActiveLog();
     if (!response.ERROR) {
@@ -64,12 +74,27 @@ function update() {
 }
 
 async function deleteFile(log) {
-    response = await deleteLog(log);
-    if (!response.ERROR) {
-        alert(`Log ${log} was deleted`);
-        location.reload();
+    if (confirm(`Do you really want to delete ${log}?`)) {
+        response = await deleteLog(log);
+        if (!response.ERROR) {
+            alert(`Log ${log} was deleted`);
+            location.reload();
+        }
+        else {
+            alert(`An error occurred: ${response.ERROR}`);
+        }
     }
-    else {
-        alert(`An error occurred: ${response.ERROR}`);
+}
+
+async function deleteAllFiles() {
+    if (confirm("Do you really want to delete all Closed and Empty logs?")) {
+        response = await deleteAll();
+        if (!response.ERROR) {
+            alert(`All Closed and Empty logs was deleted`);
+            location.reload();
+        }
+        else {
+            alert(`An error occurred: ${response.ERROR}`);
+        }
     }
 }
